BJS2-43609 [Развернись!] Исправление замечаний: - убран неоптимальный метод - метод reverse преобразован в void - создал свой класс Main для задачи
