+++ src/main/java/school/faang/moduleone/task_43500/Main.java
@@ -1,0 +1,11 @@
+package school.faang.moduleone.task_43500;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        int[] source = {3, 10, 15, 8, 12};
+        Example.reverse(source);
+        System.out.println(Arrays.toString(source));
+    }
+}
